Export tag hooks registration helper
Change-type: patch
