+++ src/features/tags/hooks.ts
@@ -1,21 +1,3 @@
-import { hooks } from '@balena/pinejs';
+import { registerTagHooks } from './validation';
 
-import { checkTagKeyValidity } from './validation';
-
-// Tag hooks
-const addTagHooks = (resource: string) => {
-	const nameProp = 'tag_key';
-	const tagHook: hooks.Hooks = {
-		POSTPARSE: ({ request }) => {
-			if (request.values[nameProp] != null) {
-				checkTagKeyValidity(request.values[nameProp]);
-			}
-		},
-	};
-
-	hooks.addPureHook('POST', 'resin', resource, tagHook);
-	hooks.addPureHook('PUT', 'resin', resource, tagHook);
-	hooks.addPureHook('PATCH', 'resin', resource, tagHook);
-};
-
-['application_tag', 'device_tag', 'release_tag'].forEach(addTagHooks);
+['application_tag', 'device_tag', 'release_tag'].forEach(registerTagHooks);
